pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Alexander Sidorov <pl4zie@gmail.com>"]
packages = [
{ include = "task_manager" }
]

[tool.poetry.dependencies]
python = "3.8.13"
django-bootstrap4 = "^22.1"
python-dotenv = "^0.20.0"
django-debug-toolbar = "^3.2.4"
psycopg2-binary = "^2.9.3"
dj-database-url = "^0.5.0"
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
pytest-django = "^4.5.2"
gunicorn = "^20.1.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
